backend/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic[email]
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

python-dotenv
pytube
youtube-transcript-api
ffmpeg-python==0.2.0
python-ffmpeg==2.0.10  # Optional: provides additional ffmpeg functionality
